StoreKPLite/requirements.txt
aiogram==3.4.1
sqlalchemy==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0
alembic==1.13.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
redis==5.0.1
cbrapi==0.1.0
aiohttp==3.9.1
httpx==0.25.2
Pillow==10.3.0
rapidfuzz>=3.0.0
openpyxl==3.1.2
reportlab==4.0.9
geoip2==4.8.0

